Move http utility file to helpers folder
